refactor: major architecture overhaul + iOS App Store prep + version upgrades
- Upgrade Compose Multiplatform 1.8.2 → 1.10.3, Kotlin 2.2.0 → 2.2.20,
  kotlinx-datetime 0.6.1 → 0.7.1, kotlin-inject 0.8.0 → 0.9.0
- Add centralized SheetManager with coroutine-based sheet results,
  replacing scattered boolean states and callback spaghetti in HomeScreen
- Split HomeScreen into InPeriodContent + NotInPeriodContent fragments
- Reorganize UI into ui/pages/ package structure
- Add in-app language switch (System/EN/ZH) with platform-specific
  LocalAppLocale implementations for Android and iOS
- Add Comfortaa custom font across all typography levels
- Improve calendar: full month names, taller rows, removed clutter dots
- Remove multi-palette feature, keep single warm theme
- Simplify HomeViewModel to only expose refresh()
- iOS App Store prep: bundle ID, privacy manifest, localized display names
- Fix kotlinx-datetime 0.7.1 breaking changes (Clock.System → kotlin.time)

+++ composeApp/src/commonMain/kotlin/com/haodong/yimalaile/domain/menstrual/MenstrualService.kt
@@ -1,6 +1,6 @@
 package com.haodong.yimalaile.domain.menstrual
 
-import kotlinx.datetime.Clock
+import kotlin.time.Clock
 import kotlinx.datetime.DateTimeUnit
 import kotlinx.datetime.LocalDate
 import kotlinx.datetime.plus
@@ -124,14 +124,14 @@ class MenstrualService(private val repository: RecordsRepository) {
     private fun averageCycleLength(records: List<MenstrualRecord>): Int? {
         val sorted = records.filter { !it.isDeleted }.sortedBy { it.startDate }
         if (sorted.size < 2) return null
-        val gaps = sorted.zipWithNext().map { (a, b) -> a.startDate.until(b.startDate, DateTimeUnit.DAY) }
+        val gaps = sorted.zipWithNext().map { (a, b) -> a.startDate.until(b.startDate, DateTimeUnit.DAY).toInt() }
         return gaps.sum() / gaps.size
     }
 
     private fun averagePeriodLength(records: List<MenstrualRecord>): Int? {
         val lengths = records
             .filter { !it.isDeleted && it.endDate != null }
-            .map { it.startDate.until(it.endDate!!, DateTimeUnit.DAY) + 1 }
+            .map { it.startDate.until(it.endDate!!, DateTimeUnit.DAY).toInt() + 1 }
         return if (lengths.isEmpty()) null else lengths.sum() / lengths.size
     }
